Refactor Appsignal Datadog Prometheus loggers
